% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransComp.R
\name{CompTransc}
\alias{CompTransc}
\title{Identify If Two Datasets Could Look at Same Transcript}
\usage{
CompTransc(transComp1, transComp2)
}
\arguments{
\item{transComp1}{a dataset containing a list of genes and their exons}

\item{transComp2}{a second dataset containing a list of genes and their exons}
}
\value{
Returns a list of genes that show overlap in transcripts. If there are none, an empty list will be returned.
}
\description{
For each common gene in dataset1 and dataset2, identify whether they could be looking at the same
transcript based on the genetic coordinates listed.
}
