% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{formattedData}
\alias{formattedData}
\title{Example BRCA2 and TP53 Dataframe}
\format{
A dataframe with 7 columns:
\describe{
\item{sample}{The sample number}
\item{ensembl_gene_id}{The ensembl gene ID}
\item{ensembl_transcript_id}{The ensembl transcript ID}
\item{ensembl_exon_id}{The ensembl exon ID}
\item{exon_chrom_start}{The exon start coordinate.}
\item{exon_chrom_end}{The exon end coordinate.}
\item{Length}{The transcript length.}
}
}
\source{
The dataset was downloaded from Ensembl
https://useast.ensembl.org/index.html.
}
\usage{
formattedData
}
\description{
A dataframe of gene and genomic coordinate data from Ensembl for BRCA2 and TP53.
This is the expected resulting dataframe when "inst/extdata/example_data.csv"
is used as the input file for function loadAndCleanData.
}
\examples{
\dontrun{
 formattedData
}
}
\references{
Dyer C Sarah et al. Ensembl 2025. Nucleic Acids Res. 2025, 53(D1):D948–D957. PMID: 39656687. doi:10.1093/nar/gkae1071
}
\keyword{datasets}
