% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransComp.R
\name{IndTransc}
\alias{IndTransc}
\title{Identify Gene Transcript Based on Genomic Coordinates}
\usage{
IndTransc(dataset1, geneID = NA, sample = NA)
}
\arguments{
\item{dataset1}{a dataframe containing the ensembl exon and transcript IDs.}

\item{geneID}{the ensembl gene ID of the target gene.}

\item{sample}{the sample number.}
}
\value{
Returns the transcript this sampled gene could have been sequenced
from or a list if there are multiple possibilities. If there are none,
it will return an empty list.
}
\description{
Identify which transcript of a gene the given genomic coordinates belong to.
These should be grouped by sequenced transcript.
This function is not available to the user.
}
