% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndCleanData.R
\name{loadAndCleanData}
\alias{loadAndCleanData}
\title{Load and Clean a Given Dataset}
\usage{
loadAndCleanData(transFile = NA)
}
\arguments{
\item{transFile}{input csv file. This file should contain genomic coordinates,
and either the ensembl gene ID or both a gene identifier and source.
It can also include a column detailing which rows were sequenced as part of the same sample,
which can be useful for some downstream tasks.}
}
\value{
Returns a cleaned dataset.
}
\description{
Given a file with the specified inputs, load and clean the dataset
so that it contains standardized ensembl gene IDs, exon IDs, and transcript IDs.
}
\examples{
filepath <- system.file("extdata", "example_data.csv", package = "transComp")
results <- loadAndCleanData(filepath)
results

}
\references{
BioMart and Bioconductor: a powerful link between biological databases and microarray data analysis.
Steffen Durinck, Yves Moreau, Arek Kasprzyk, Sean Davis, Bart De Moor, Alvis Brazma and Wolfgang Huber,
Bioinformatics 21, 3439-3440 (2005).

Wickham H (2025). stringr: Simple, Consistent Wrappers for Common String Operations.
doi:10.32614/CRAN.package.stringr \url{https://doi.org/10.32614/CRAN.package.stringr}, R package version 1.5.2,

R Core Team (2025). R: A Language and Environment for Statistical Computing.
R Foundation for Statistical Computing, Vienna, Austria. \url{https://www.R-project.org/}.

Bache S, Wickham H (2025). magrittr: A Forward-Pipe Operator for R. doi:10.32614/CRAN.package.magrittr
\url{https://doi.org/10.32614/CRAN.package.magrittr}, R package version 2.0.4,
\url{https://CRAN.R-project.org/package=magrittr}.
}
