% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndCleanData.R
\name{loadCleanData}
\alias{loadCleanData}
\title{Load and Clean a Given Dataset}
\usage{
loadCleanData(transFile = NA)
}
\arguments{
\item{transFile}{input csv file. This file should contain genomic coordinates, and either the ensembl gene ID or both a gene identifier and source. It can also include a column detailing which rows were sequenced as part of the same sample, which can be useful for some downstream tasks.}
}
\value{
Returns a cleaned dataset.
}
\description{
Given a file, load and clean the dataset so that it contains standardized ensembl
gene IDs, exon IDs, and transcript IDs.
}
