% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCanTrans.R
\name{plotTransCan}
\alias{plotTransCan}
\title{Plot Canonical Transcript Composition.}
\usage{
plotTransCan(transData, datasetName = NA)
}
\arguments{
\item{transData}{a dataframe containing ensembl gene, transcript, and exon IDs.}

\item{datasetName}{(optional) name of dataset provided to include in figure.}
}
\value{
Returns a plot of transcript composition.
}
\description{
Plot the canonical transcript composition of this dataset based on transcripts
associated with the exon composition associated with each gene. This function
will plot the relative count of canonical transcripts compared to non-canonical transcripts
found for this dataset.
}
\examples{
canPlot <- plotTransCan(formattedData, "test data")
canPlot

}
\references{
BioMart and Bioconductor: a powerful link between biological databases and microarray data analysis.
Steffen Durinck, Yves Moreau, Arek Kasprzyk, Sean Davis, Bart De Moor,
Alvis Brazma and Wolfgang Huber, Bioinformatics 21, 3439-3440
(2005).

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.

Wickham H, François R, Henry L, Müller K, Vaughan D (2023).
\emph{dplyr: A Grammar of Data Manipulation}. doi:10.32614/CRAN.package.dplyr
\url{https://doi.org/10.32614/CRAN.package.dplyr}, R package version 1.1.4,
\url{https://CRAN.R-project.org/package=dplyr}.
}
