% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCanTrans.R
\name{plotTransComp}
\alias{plotTransComp}
\title{Plot Canonical Transcript Composition.}
\usage{
plotTransComp(transData, datasetName)
}
\arguments{
\item{transData}{a dataframe containing ensembl gene, transcript, and exon IDs.}

\item{datasetName}{(optional) name of dataset provided to include in figure.}
}
\value{
Returns a plot of transcript composition.
}
\description{
Plot the canonical transcript composition of this dataset based on transcripts
associated with the exon composition associated with each gene.
}
