% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompExons.R
\name{SameTransc}
\alias{SameTransc}
\title{Identify If Two Datasets Could Look at Same Transcript}
\usage{
SameTransc(dataset1, dataset2, geneID)
}
\arguments{
\item{dataset1}{first dataset to compare. This dataset should contain the genomic coordinates for each sample.}

\item{dataset2}{second dataset to compare. This dataset should contain the genomic coordinates for each sample.}

\item{geneID}{the ensembl gene ID of the target gene.}
}
\value{
Returns a list of genes that both datasets could be looking at. If there are none, an empty list will be returned.
}
\description{
Identify whether two datasets are looking at the same transcripts for a gene.
}
