% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{secondExample}
\alias{secondExample}
\title{Example BRCA2 Dataframe}
\format{
A dataframe with two columns:
\describe{
\item{sample}{The sample number}
\item{ensembl_gene_id}{The ensembl gene ID}
\item{ensembl_transcript_id}{The ensembl transcript ID}
\item{ensembl_exon_id}{The ensembl exon ID}
\item{exon_chrom_start}{The exon start coordinate.}
\item{exon_chrom_end}{The exon end coordinate.}
\item{Length}{The transcript length.}
}
}
\source{
The dataset was downloaded from Ensembl
https://useast.ensembl.org/index.html
}
\usage{
secondExample
}
\description{
A dataframe of gene and genomic coordinate data from Ensembl for BRCA2.
This is the expected resulting dataframe when "inst/extdata/example_data2.csv"
is used as the input file for function loadAndCleanData. This is used for the example
for function "CompTransc".
}
\examples{
\dontrun{
 secondExample
}
}
\references{
Dyer C Sarah et al. Ensembl 2025. Nucleic Acids Res. 2025,
53(D1):D948–D957. PMID: 39656687. doi:10.1093/nar/gkae1071
}
\keyword{datasets}
