% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transComp-package.R
\docType{package}
\name{transComp-package}
\alias{transComp}
\alias{transComp-package}
\title{transComp: Determine and Compare Transcript Composition of Datasets}
\description{
Determine transcript composition of a dataset or identify whether two transcripts are the same given genomic coordinates or ensembl labels using the hg38 reference genome.
}
\author{
\strong{Maintainer}: Sabrina Xi \email{sabrina.xi@mail.utoronto.ca}

}
\keyword{internal}
