% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransComp.R
\name{TransComp}
\alias{TransComp}
\title{Identify Gene Transcript Based on Genomic Coordinates}
\usage{
TransComp(transData)
}
\arguments{
\item{transData}{a dataframe containing the ensembl transcript, exon, and gene IDs. It should also include a column identifying grouped exons.}
}
\value{
Returns a dataframe with the possible transcript of each gene read.
}
\description{
Identify the transcript composition of a given dataset, given multiple sequencing reads of a gene.
}
