% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanTrans.R
\name{CanTrans}
\alias{CanTrans}
\title{Identify a Dataset's Canonical Transcript Composition}
\usage{
CanTrans(transData)
}
\arguments{
\item{transData}{dataframe containing gene and exon data. It should also contain a column identifying exons that were sequenced from the same transcript.}
}
\value{
Returns a list of genes that only contains exons found on the canonical transcript.
}
\description{
Identify the composition of canonical transcripts of a dataset from its exon composition.
}
\examples{
canonical_transcripts <- CanTrans(formattedData)
canonical_transcripts

}
\references{
BioMart and Bioconductor: a powerful link between biological databases and microarray data analysis. Steffen Durinck,
Yves Moreau, Arek Kasprzyk, Sean Davis, Bart De Moor, Alvis Brazma and Wolfgang Huber, Bioinformatics 21, 3439-3440
(2005).
}
