% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompExons.R
\name{CompTransc}
\alias{CompTransc}
\title{Identify If Two Datasets Could Look at Same Transcript}
\usage{
CompTransc(transComp1, transComp2)
}
\arguments{
\item{transComp1}{a dataset containing a list of genes and their possible transcripts.}

\item{transComp2}{a second dataset containing a list of genes and their possible transcripts.}
}
\value{
Returns a list of genes that show overlap in transcripts. If there are none, an empty list will be returned.
}
\description{
Identify whether two datasets are looking at the same transcripts for each gene.
}
