% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransComp.R
\name{IndTransc}
\alias{IndTransc}
\title{Identify Gene Transcript Based on Genomic Coordinates}
\usage{
IndTransc(dataset1, geneID)
}
\arguments{
\item{dataset1}{a dataframe containing the ensembl exon and transcript IDs.}

\item{geneID}{the ensembl gene ID of the target gene.}
}
\value{
Returns a list of possible transcripts this gene could be. If there are none, the list will be empty.
}
\description{
Identify which transcript of a gene the given genomic coordinates belong to.
}
