% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTransComp.R
\name{plotTransComp}
\alias{plotTransComp}
\title{Plot Transcript Composition.}
\usage{
plotTransComp(transData, datasetName)
}
\arguments{
\item{transData}{a dataframe containing ensembl gene, transcript, and exon IDs.}

\item{datasetName}{(optional) name of dataset provided to include in figure.}
}
\value{
Returns a plot of transcript composition.
}
\description{
Plot the transcript composition of this dataset. This function will plot the
count of each transcript associated with the dataset for each, using the
result from TransComp based on genomic coordinates given.
}
\examples{
plot <- plotTransComp(formattedData, "test data")
plot

}
\references{
BioMart and Bioconductor: a powerful link between biological databases and microarray data analysis. Steffen Durinck,
Yves Moreau, Arek Kasprzyk, Sean Davis, Bart De Moor, Alvis Brazma and Wolfgang Huber, Bioinformatics 21, 3439-3440
(2005).

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.

Wickham H, François R, Henry L, Müller K, Vaughan D (2023). \emph{dplyr: A Grammar of Data Manipulation}.
doi:10.32614/CRAN.package.dplyr \url{https://doi.org/10.32614/CRAN.package.dplyr}, R package version 1.1.4,
\url{https://CRAN.R-project.org/package=dplyr}.
}
