% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransComp.R
\name{TransComp}
\alias{TransComp}
\title{Identify Gene Transcripts Based on Genomic Coordinates}
\usage{
TransComp(transData)
}
\arguments{
\item{transData}{a dataframe containing the ensembl transcript, exon, and gene IDs.
It should also include a column identifying grouped exons.}
}
\value{
Returns a dataframe with the possible transcript of each gene read.
}
\description{
Identify the transcript composition of a given dataset, given multiple sequencing reads of a gene.
}
\examples{
canonical_transcripts <- TransComp(formattedData)
canonical_transcripts

}
\references{
Wickham H, François R, Henry L, Müller K, Vaughan D (2023). \emph{dplyr: A Grammar of Data Manipulation}.
doi:10.32614/CRAN.package.dplyr \url{https://doi.org/10.32614/CRAN.package.dplyr}, R package version 1.1.4,
\url{https://CRAN.R-project.org/package=dplyr}.
}
